refactor(sdk): rename ai_hub/manifest to ai_hub/dto
